# specs/003-calc-history-persistence/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Calculator History API
  description: |
    Endpoints for retrieving and clearing an authenticated user's calculation history.
    All endpoints require an active authenticated session (cookie-based, managed by Spring Security).
    Unauthenticated requests receive `401 Unauthorized` with a JSON error body.
  version: "1.0.0"

servers:
  - url: http://localhost:8080
    description: Local development server

security:
  - cookieAuth: []

paths:
  /api/history:
    get:
      operationId: getCalculationHistory
      summary: Get the current user's calculation history
      description: |
        Returns the most recent 20 calculation entries for the authenticated user,
        ordered newest first. Returns an empty array when no history exists.
        Only the calling user's entries are returned; no cross-user access is possible.
      tags:
        - History
      responses:
        "200":
          description: List of calculation history entries (may be empty)
          content:
            application/json:
              schema:
                type: array
                maxItems: 20
                items:
                  $ref: "#/components/schemas/HistoryEntry"
              examples:
                withEntries:
                  summary: User with two history entries
                  value:
                    - operandA: 10.0
                      operandB: 3.0
                      operation: "DIVIDE"
                      result: 3.3333333333333335
                      calculatedAt: "2026-05-15T10:32:01.456Z"
                    - operandA: 5.0
                      operandB: 4.0
                      operation: "ADD"
                      result: 9.0
                      calculatedAt: "2026-05-15T10:31:45.123Z"
                empty:
                  summary: User with no history
                  value: []
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Authentication required."

    delete:
      operationId: clearCalculationHistory
      summary: Clear the current user's calculation history
      description: |
        Permanently deletes all calculation history entries for the authenticated user.
        Returns `204 No Content` on success. The operation is idempotent — calling it
        when no history exists also returns `204`.
        Only the calling user's entries are deleted; other users' histories are unaffected.

        **CSRF note**: This endpoint is under `/api/**` which is exempt from CSRF
        validation in `SecurityConfig`. No `X-XSRF-TOKEN` header is required.
      tags:
        - History
      responses:
        "204":
          description: History cleared successfully (or was already empty)
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Authentication required."

components:
  schemas:
    HistoryEntry:
      type: object
      description: A single saved calculation belonging to the authenticated user
      required:
        - operandA
        - operandB
        - operation
        - result
        - calculatedAt
      properties:
        operandA:
          type: number
          format: double
          description: Left-hand operand
          example: 10.0
        operandB:
          type: number
          format: double
          description: Right-hand operand
          example: 3.0
        operation:
          type: string
          description: Operation name (matches the `Operation` enum)
          enum:
            - ADD
            - SUBTRACT
            - MULTIPLY
            - DIVIDE
          example: "DIVIDE"
        result:
          type: number
          format: double
          description: Computed result
          example: 3.3333333333333335
        calculatedAt:
          type: string
          format: date-time
          description: UTC timestamp of when the calculation was performed (ISO-8601)
          example: "2026-05-15T10:32:01.456Z"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Authentication required."

  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: |
        Spring Security session cookie. Obtained by POSTing valid credentials to `/login`.
        The `XSRF-TOKEN` cookie is also set but is not required for `/api/**` endpoints
        (they are CSRF-exempt per `SecurityConfig`).